refactor: Replace inline status color logic with StatusIndicator component for ServerItem and SnapshotItem
